headless_lms_models

Module certificate_configurations

Structs

CertificateConfiguration

CertificateConfigurationAndRequirements

DatabaseCertificateConfiguration

DatabaseCertificateConfigurationInner 🔒

Enums

CertificateTextAnchor

How text should be positioned relative to the given coordinates. See https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/text-anchor.

PaperSize

Functions

delete

get_by_id

get_default_certificate_configurations_and_requirements_by_course

Returns all default certificate configurations for a course, one per course module. A default certificate configuration requires only one course module.

get_default_configuration_by_course_module

get_first_configuration_id

insert

update