headless_lms_models

Module chapters

Structs

Chapter

ChapterAvailability

ChapterExerciseDeadlineOverrideSummary

ChapterInfo

ChapterLockPreview

ChapterPagesWithExercises

ChapterUpdate

ChapterWithStatus

CourseUserInfo

DatabaseChapter

NewChapter

UnreturnedExercise

UserChapterProgress

UserCourseInstanceChapterProgress

Enums

ChapterStatus

Functions

course_chapters

course_instance_chapters

delete_chapter

exercise_deadline_overrides_by_chapter_for_course

fetch_chapter_availability

fetch_course_users

fetch_user_chapter_progress

get_chapter

get_chapter_by_page_id

get_chapter_info_by_page_metadata

get_chapter_lock_preview

get_course_id

get_for_module

get_previous_chapters_in_module

get_user_course_instance_chapter_progress

insert

is_open

Checks the opens_at field for the chapter and compares it to the current time. If null, the chapter is always open.

move_chapter_exercises_to_manual_review

set_front_page

set_module

set_opens_at

unlock_first_chapters_for_user

Unlocks the first chapter(s) with exercises in the base module (order_number == 0) for a user. Also unlocks any chapters without exercises that come before the first chapter with exercises.

unlock_next_chapters_for_user

Unlocks the next chapter(s) for a user after they complete a chapter. If the completed chapter is the last in a base module (order_number == 0), unlocks the first chapter of all additional modules (order_number != 0). Otherwise, unlocks the next chapter in the same module. Note: If a module has no chapters with exercises, all chapters in that module will be unlocked. This is intentional to allow progression through content-only chapters.

update_chapter

update_chapter_image_path