Struct headless_lms_models::error::ModelError

pub struct ModelError {
    error_type: ModelErrorType,
    message: String,
    source: Option<Error>,
    span_trace: SpanTrace,
    backtrace: Backtrace,
}

Error type used by all models. Used as the error type in ModelError, which is used by all the controllers in the application.

All the information in the error is meant to be seen by the user. The type of error is determined by the ModelErrorType enum, which is stored inside this struct.

Examples

Usage without source error

if erroneous_condition {
    return Err(ModelError::new(
        ModelErrorType::PreconditionFailed,
        "The user has not enrolled to this course".to_string(),
        None,
    ));
}

Usage with a source error

Used when calling a function that returns an error that cannot be automatically converted to an ModelError. (See impl From<X> implementations on this struct.)

some_function_returning_an_error().map_err(|original_error| {
    ModelError::new(
        ModelErrorType::Generic,
        "Everything went wrong".to_string(),
        Some(original_error.into()),
    )
})?;

Fields
error_type: ModelErrorTypemessage: Stringsource: Option<Error>

Original error that caused this error.

span_trace: SpanTrace

A trace of tokio tracing spans, generated automatically when the error is generated.

backtrace: Backtrace

Stack trace, generated automatically when the error is created.