headless_lms_models

Module exercise_reset_logs

Structs

ExerciseResetLog

Functions

get_exercise_reset_logs_for_user

log_exercise_reset

Adds a log entry for reset exercises for a user

user_should_see_reset_message_for_exercise

Check if the user’s exercise has been reset and no new submissions have been made since.