Module headless_lms_models::exercise_task_gradings

Structs

ExerciseTaskGrading

ExerciseTaskGradingRequest

ExerciseTaskGradingResult

Enums

UserPointsUpdateStrategy

Functions

get_all_gradings_by_exercise_slide_submission_id

get_by_exercise_task_submission_id

get_by_id

get_course_id

get_course_or_exam_id

get_for_student

Fetches the grading for the student, but hides the result in some circumstances. For example, for an ongoing exam.

get_new_and_old_exercise_task_gradings_by_regrading_id

get_point_update_strategy_from_gradings

For now gets this information from some task submission in a slide submission.

get_total_score_given_for_exercise_slide_submission

get_user_exercise_task_gradings_by_module_and_exercise_type

grade_submission

insert

new_grading

set_grading_progress

update_grading