headless_lms_models

Module exercises

Structs

CourseMaterialExercise

Exercise

ExerciseGradingStatus

ExerciseStatus

ExerciseStatusSummaryForUser

Enums

ActivityProgress

Indicates what is the user’s completion status for a exercise.

GradingProgress

Tells what’s the status of the grading progress for a user and exercise.

Functions

collect_user_ids_and_exercise_ids_for_reset

Collects user_ids and related exercise_ids according to given filters

delete_exercises_by_page_id

determine_can_post_submission 🔒

get_all_exercise_statuses_by_user_id_and_course_id

get_by_id

get_course_material_exercise

get_course_or_exam_id

get_exercise_by_id

get_exercise_submissions_and_status_by_course_instance_id

get_exercises_by_chapter_id

get_exercises_by_chapter_ids

get_exercises_by_course_id

get_exercises_by_exam_id

get_exercises_by_module_containing_exercise_type

get_exercises_by_page_id

get_or_select_exercise_slide

insert

reset_exercises_for_selected_users

Resets all related tables for selected users and related exercises

set_exercise_to_use_exercise_specific_peer_or_self_review_config