headless_lms_models

Module library

Modules

content_management

copying

course_instances

General functionality related to course instances

course_stats

custom_view_exercises

global_stats

grading

Collection of functions used for processing and evaluating user submissions for exercises.

page_visit_stats

Functions for precomputing anonymous aggregate statistics about page visits.

peer_or_self_reviewing

progressing

regrading

user_exercise_state_updater

Always update the user_exercise_state table though this module

Structs

ParseTimeGranularityError

Enums

TimeGranularity