headless_lms_models::library
Module global_stats

Structs

CourseCompletionStats

DomainCompletionStats

GlobalCourseModuleStatEntry

GlobalStatEntry

Functions

get_completion_stats_by_email_domain

Produces a summary of course completions grouped by user’s email domain.

get_course_completion_stats_for_email_domain

Gets course completion statistics for a specific email domain.

get_course_module_stats_by_completions_registered_to_study_registry

get_number_of_people_completed_a_course

get_number_of_people_done_at_least_one_exercise

get_number_of_people_registered_completion_to_study_registry

get_number_of_people_started_course