Function headless_lms_models::library::grading::update_user_exercise_slide_state_and_user_exercise_state

async fn update_user_exercise_slide_state_and_user_exercise_state(
    conn: &mut PgConnection,
    user_exercise_slide_state: UserExerciseSlideState,
    user_points_update_strategy: UserPointsUpdateStrategy,
) -> ModelResult<UserExerciseState>

Updates the user exercise state starting from a slide state, and propagates the update up to the whole user exercise state.