headless_lms_models::library

Module peer_or_self_reviewing

Structs

CourseMaterialPeerOrSelfReviewData

CourseMaterialPeerOrSelfReviewDataAnswerToReview

CourseMaterialPeerOrSelfReviewQuestionAnswer

CourseMaterialPeerOrSelfReviewSubmission

Constants

MAX_PEER_REVIEW_CANDIDATES 🔒

Functions

create_peer_or_self_review_submission_for_user

get_course_material_peer_or_self_review_data 🔒

select_own_submission_for_self_review

Selects a user’s own submission to be self-reviewed. Works similarly to try_to_select_exercise_slide_submission_for_peer_review but selects the user’s latest submission.

start_peer_or_self_review_for_user

Starts peer review state for the student for this exercise.

try_to_select_exercise_slide_submission_for_peer_review

Tries to select a submission for user to peer review.

try_to_select_peer_review_candidate_from_queue 🔒

try_to_select_peer_review_candidate_from_queue_impl 🔒

Returns a tuple of the exercise slide submission id and a boolean indicating if the submission needs peer review.

update_peer_review_queue_reviews_received

update_peer_review_receiver_exercise_status 🔒

validate_and_sanitize_peer_review_submission_answers 🔒

Filters submitted peer review answers to those that are part of the peer review.