Module headless_lms_models::library::peer_reviewing

Structs

CourseMaterialPeerReviewData

CourseMaterialPeerReviewDataAnswerToReview

CourseMaterialPeerReviewQuestionAnswer

CourseMaterialPeerReviewSubmission

Constants

MAX_PEER_REVIEW_CANDIDATES 🔒

Functions

create_peer_review_submission_for_user

get_course_material_peer_review_data 🔒

start_peer_review_for_user

Starts peer review state for the student for this exercise.

try_to_select_exercise_slide_submission_for_peer_review

Tries to select a submission for user to peer review.

try_to_select_peer_review_candidate_from_queue 🔒

update_peer_review_giver_exercise_progress 🔒

Creates or updates submitter’s exercise state and peer review queue entry.

update_peer_review_receiver_exercise_status 🔒

validate_and_sanitize_peer_review_submission_answers 🔒

Filters submitted peer review answers to those that are part of the peer review.