Module headless_lms_models::library::progressing

Structs

CompletionRegistrationLink

CourseInstanceCompletionSummary

ManualCompletionPreview

ManualCompletionPreviewUser

TeacherManualCompletion

TeacherManualCompletionRequest

UserCompletionInformation

UserCourseModuleCompletion

UserModuleCompletionStatus

UserWithModuleCompletions

Statics

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

CALLSITE 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

META 🔒

Functions

add_manual_completions

create_automatic_course_module_completion_if_eligible 🔒

Creates completion for the user if eligible and previous one doesn’t exist. Returns a boolean indicating whether a completion exists after calling this function.

get_completion_registration_link_and_save_attempt

get_course_instance_completion_summary

get_manual_completion_result_preview

Gets a preview of changes that will occur to completions with the given manual completion data.

get_user_completion_information

get_user_module_completion_statuses_for_course_instance

Gets course modules with user’s completion status for the given instance.

process_all_course_completions

Goes through all course instances on a course and grants completions to users on those courses where eligible.

process_all_course_instance_completions

Goes through all users on the course instance and grants them completions where eligible.

update_automatic_completion_status_and_grant_if_eligible

Checks whether the course module can be completed automatically and creates an entry for completion if the user meets the criteria. Also re-checks module completion prerequisites if the module is completed.

update_module_completion_prerequisite_statuses_for_user 🔒

Fetches all course module completions for the given user on the given course and updates the prerequisite module completion statuses for any completions that are missing them.

user_can_take_exam

Checks whether the student can partake in an exam.

user_has_passed_exam_for_the_course 🔒

Returns true if there is at least one exam associated with the course, that has ended and the user has received enough points from it.

user_is_eligible_for_automatic_completion 🔒

user_passes_automatic_completion_exercise_tresholds 🔒