headless_lms_models
Module offered_answers_to_peer_review_temporary

Functions

delete_expired_records 🔒

Deletes entries older than 1 hour – for keeping the table small and fast

delete_saved_submissions_for_user

For clearing the table after the user has given a peer review so that they can receive a new submission to be reviewed

save_given_exercise_slide_submission

Returns an exercise slide submission id that has been given to be reviewed by the student within the hour.

try_to_restore_previously_given_exercise_slide_submission

Returns an exercise slide submission id that has been given to be reviewed by the student within the hour. Does not return submissions that no longer need peer review.