headless_lms_models::page_language_groups

Function get_all_pages_in_page_language_group_mapping

pub async fn get_all_pages_in_page_language_group_mapping(
    conn: &mut PgConnection,
    page_id: Uuid,
) -> ModelResult<HashMap<CourseOrExamId, PageLanguageGroupNavigationInfo>>

Can be used to find the same page in all the different language versions of the course. Returns a mappgig from course_id or exam_id to the page language group navigation info.