headless_lms_models
Module peer_or_self_review_configs

Structs

CmsPeerOrSelfReviewConfig

CmsPeerOrSelfReviewConfiguration

CourseMaterialPeerOrSelfReviewConfig

Like PeerOrSelfReviewConfig but only the fields it’s fine to show to all users.

PeerOrSelfReviewConfig

Enums

PeerReviewProcessingStrategy

Determines how we will treat the answer being peer reviewed once it has received enough reviews and the student has given enough peer reviews.

Functions

delete

delete_peer_reviews_by_exrcise_ids

get_by_exercise_id

Usually you want to use get_by_exercise_or_course_id instead of this one.

get_by_exercise_or_course_id

Returns the correct peer review config depending on exercise.use_course_default_peer_or_self_review_config.

get_by_id

get_cms_peer_review_by_id

get_course_default_cms_peer_review

get_course_material_peer_or_self_review_data

get_default_for_course_by_course_id

get_peer_reviews_by_page_id

insert

upsert_course_default_cms_peer_review_and_questions

upsert_with_id