headless_lms_models
Module peer_review_queue_entries

Structs

PeerReviewQueueEntry

Functions

delete_by_receiving_peer_reviews_exercise_slide_submission_id

get_all_by_user_and_course_instance_ids

get_all_that_need_peer_reviews_by_exercise_id

Gets multiple records of PeerReviewQueueEntry that still require more peer reviews, ordered by peer review priority.

get_any_including_not_needing_review

Gets multiple records of PeerReviewQueueEntry ordered by newest queue entries first. Also returns entries that don’t need peer review.

get_by_id

get_by_receiving_peer_reviews_exercise_slide_submission_id

get_by_user_and_exercise_and_course_instance_ids

get_entries_that_need_reviews_and_are_older_than

get_entries_that_need_reviews_and_are_older_than_with_exercise_id

get_entries_that_need_teacher_review_and_are_older_than_with_course_instance_id

Returns entries that have been waiting for teacher to review them for more than timestamp. The teacher is supposed to review an answer when the user_exercise_states reviewing_stage is WaitingForManualGrading.

get_many_that_need_peer_reviews_by_exercise_id_and_review_priority

Gets multiple records of PeerReviewQueueEntry that still require more peer reviews, ordered by peer review priority.

increment_peer_review_priority

insert

remove_from_queue 🔒

remove_from_queue_and_add_to_manual_review

remove_from_queue_and_give_full_points

remove_queue_entries_for_unusual_reason

try_to_get_all_by_user_and_course_instance_ids

try_to_get_by_user_and_exercise_and_course_instance_ids

update

update_received_enough_peer_reviews

upsert_peer_review_priority

Inserts or updates the queue entry indexed by user_id, exercise_id and course_instance_id.