Module headless_lms_models::peer_reviews

Structs

PeerReview

Enums

PeerReviewAcceptingStrategy

Determines how we will treat the answer being peer reviewed once it has received enough reviews and the student has given enough peer reviews.

Functions

delete

get_by_exercise_or_course_id

get_by_id

get_course_material_peer_review_data

get_default_for_course_by_course_id

insert

insert_with_id

try_to_get_by_exercise_id