headless_lms_models::prelude::chrono::prelude

Struct Date

pub struct Date<Tz>
where
    Tz: TimeZone,
{
    date: NaiveDate,
    offset: <Tz as TimeZone>::Offset,
}

👎Deprecated since 0.4.23: Use NaiveDate or DateTime<Tz> instead

ISO 8601 calendar date with time zone.

You almost certainly want to be using a NaiveDate instead of this type.

This type primarily exists to aid in the construction of DateTimes that have a timezone by way of the TimeZone datelike constructors (e.g. TimeZone::ymd).

This type should be considered ambiguous at best, due to the inherent lack of precision required for the time zone resolution.

There are some guarantees on the usage of Date<Tz>:

If properly constructed via TimeZone::ymd and others without an error, the corresponding local date should exist for at least a moment. (It may still have a gap from the offset changes.)

The TimeZone is free to assign any Offset to the local date, as long as that offset did occur in given day.

For example, if 2015-03-08T01:59-08:00 is followed by 2015-03-08T03:00-07:00, it may produce either 2015-03-08-08:00 or 2015-03-08-07:00 but not 2015-03-08+00:00 and others.

Once constructed as a full DateTime, DateTime::date and other associated methods should return those for the original Date. For example, if dt = tz.ymd_opt(y,m,d).unwrap().hms(h,n,s) were valid, dt.date() == tz.ymd_opt(y,m,d).unwrap().

The date is timezone-agnostic up to one day (i.e. practically always), so the local date and UTC date should be equal for most cases even though the raw calculation between NaiveDate and TimeDelta may not.

Fields
date: NaiveDate
👎Deprecated since 0.4.23: Use NaiveDate or DateTime<Tz> instead
offset: <Tz as TimeZone>::Offset
👎Deprecated since 0.4.23: Use NaiveDate or DateTime<Tz> instead

Implementations

impl<Tz> Date<Tz>
where Tz: TimeZone,

pub const MIN_UTC: Date<Utc>

The minimum possible Date.

pub const MAX_UTC: Date<Utc>

The maximum possible Date.

pub fn from_utc(date: NaiveDate, offset: <Tz as TimeZone>::Offset) -> Date<Tz>

Makes a new Date with given UTC date and offset. The local date should be constructed via the TimeZone trait.

pub fn and_time(&self, time: NaiveTime) -> Option<DateTime<Tz>>

Makes a new DateTime from the current date and given NaiveTime. The offset in the current date is preserved.

Returns None on invalid datetime.

pub fn and_hms(&self, hour: u32, min: u32, sec: u32) -> DateTime<Tz>

👎Deprecated since 0.4.23: Use and_hms_opt() instead

Makes a new DateTime from the current date, hour, minute and second. The offset in the current date is preserved.

Panics on invalid hour, minute and/or second.

pub fn and_hms_opt(&self, hour: u32, min: u32, sec: u32) -> Option<DateTime<Tz>>

Makes a new DateTime from the current date, hour, minute and second. The offset in the current date is preserved.

Returns None on invalid hour, minute and/or second.

pub fn and_hms_milli( &self, hour: u32, min: u32, sec: u32, milli: u32, ) -> DateTime<Tz>

👎Deprecated since 0.4.23: Use and_hms_milli_opt() instead

Makes a new DateTime from the current date, hour, minute, second and millisecond. The millisecond part can exceed 1,000 in order to represent the leap second. The offset in the current date is preserved.

Panics on invalid hour, minute, second and/or millisecond.

pub fn and_hms_milli_opt( &self, hour: u32, min: u32, sec: u32, milli: u32, ) -> Option<DateTime<Tz>>

Makes a new DateTime from the current date, hour, minute, second and millisecond. The millisecond part can exceed 1,000 in order to represent the leap second. The offset in the current date is preserved.

Returns None on invalid hour, minute, second and/or millisecond.

pub fn and_hms_micro( &self, hour: u32, min: u32, sec: u32, micro: u32, ) -> DateTime<Tz>

👎Deprecated since 0.4.23: Use and_hms_micro_opt() instead

Makes a new DateTime from the current date, hour, minute, second and microsecond. The microsecond part can exceed 1,000,000 in order to represent the leap second. The offset in the current date is preserved.

Panics on invalid hour, minute, second and/or microsecond.

pub fn and_hms_micro_opt( &self, hour: u32, min: u32, sec: u32, micro: u32, ) -> Option<DateTime<Tz>>

Makes a new DateTime from the current date, hour, minute, second and microsecond. The microsecond part can exceed 1,000,000 in order to represent the leap second. The offset in the current date is preserved.

Returns None on invalid hour, minute, second and/or microsecond.

pub fn and_hms_nano( &self, hour: u32, min: u32, sec: u32, nano: u32, ) -> DateTime<Tz>

👎Deprecated since 0.4.23: Use and_hms_nano_opt() instead

Makes a new DateTime from the current date, hour, minute, second and nanosecond. The nanosecond part can exceed 1,000,000,000 in order to represent the leap second. The offset in the current date is preserved.

Panics on invalid hour, minute, second and/or nanosecond.

pub fn and_hms_nano_opt( &self, hour: u32, min: u32, sec: u32, nano: u32, ) -> Option<DateTime<Tz>>

Makes a new DateTime from the current date, hour, minute, second and nanosecond. The nanosecond part can exceed 1,000,000,000 in order to represent the leap second. The offset in the current date is preserved.

Returns None on invalid hour, minute, second and/or nanosecond.

pub fn succ(&self) -> Date<Tz>

👎Deprecated since 0.4.23: Use succ_opt() instead

Makes a new Date for the next date.

Panics when self is the last representable date.

pub fn succ_opt(&self) -> Option<Date<Tz>>

Makes a new Date for the next date.

Returns None when self is the last representable date.

pub fn pred(&self) -> Date<Tz>

👎Deprecated since 0.4.23: Use pred_opt() instead

Makes a new Date for the prior date.

Panics when self is the first representable date.

pub fn pred_opt(&self) -> Option<Date<Tz>>

Makes a new Date for the prior date.

Returns None when self is the first representable date.

pub fn offset(&self) -> &<Tz as TimeZone>::Offset

Retrieves an associated offset from UTC.

pub fn timezone(&self) -> Tz

Retrieves an associated time zone.

pub fn with_timezone<Tz2>(&self, tz: &Tz2) -> Date<Tz2>
where Tz2: TimeZone,

Changes the associated time zone. This does not change the actual Date (but will change the string representation).

pub fn checked_add_signed(self, rhs: TimeDelta) -> Option<Date<Tz>>

Adds given TimeDelta to the current date.

Returns None when it will result in overflow.

pub fn checked_sub_signed(self, rhs: TimeDelta) -> Option<Date<Tz>>

Subtracts given TimeDelta from the current date.

Returns None when it will result in overflow.

pub fn signed_duration_since<Tz2>(self, rhs: Date<Tz2>) -> TimeDelta
where Tz2: TimeZone,

Subtracts another Date from the current date. Returns a TimeDelta of integral numbers.

This does not overflow or underflow at all, as all possible output fits in the range of TimeDelta.

pub fn naive_utc(&self) -> NaiveDate

Returns a view to the naive UTC date.

pub fn naive_local(&self) -> NaiveDate

Returns a view to the naive local date.

This is technically the same as naive_utc because the offset is restricted to never exceed one day, but provided for the consistency.

pub fn years_since(&self, base: Date<Tz>) -> Option<u32>

Returns the number of whole years from the given base until self.

impl<Tz> Date<Tz>
where Tz: TimeZone, <Tz as TimeZone>::Offset: Display,

pub fn format_with_items<'a, I, B>(&self, items: I) -> DelayedFormat<I>
where I: Iterator<Item = B> + Clone, B: Borrow<Item<'a>>,

Formats the date with the specified formatting items.

pub fn format<'a>(&self, fmt: &'a str) -> DelayedFormat<StrftimeItems<'a>>

Formats the date with the specified format string. See the crate::format::strftime module on the supported escape sequences.