headless_lms_models::prelude::futures::io

Struct LineWriter

pub struct LineWriter<W>
where
    W: AsyncWrite,
{
    buf_writer: BufWriter<W>,
}

Wrap a writer, like BufWriter does, but prioritizes buffering lines

This was written based on std::io::LineWriter which goes into further details explaining the code.

Buffering is actually done using BufWriter. This class will leverage BufWriter to write on-each-line.

Fields
buf_writer: BufWriter<W>
Implementations

impl<W> LineWriter<W>
where W: AsyncWrite,

pub fn new(inner: W) -> LineWriter<W>

Create a new LineWriter with default buffer capacity. The default is currently 1KB which was taken from std::io::LineWriter

pub fn with_capacity(capacity: usize, inner: W) -> LineWriter<W>

Creates a new LineWriter with the specified buffer capacity.

pub fn buffer(&self) -> &[u8]

Returns a reference to buf_writer’s internally buffered data.

pub fn get_ref(&self) -> &W

Acquires a reference to the underlying sink or stream that this combinator is pulling from.