headless_lms_models

Module re_exports

Re-exports

pub use crate::CourseOrExamId;

pub use crate::HttpErrorType;

pub use crate::ModelError;

pub use crate::ModelErrorType;

pub use crate::ModelResult;

pub use crate::PKeyPolicy;

pub use crate::error::TryToOptional;

Modules

anyhow

github crates-io docs-rs

chrono

Chrono: Date and Time for Rust

futures

Abstractions for asynchronous programming.

serde_json

Serde JSON

tokio

A runtime for writing reliable network applications without compromising speed.

Macros

debug

Constructs an event at the debug level.

error

Constructs an event at the error level.

info

Constructs an event at the info level.

trace

Constructs an event at the trace level.

warn

Constructs an event at the warn level.

Structs

ApplicationConfiguration

AzureBlobStorageConfiguration

AzureChatbotConfiguration

AzureConfiguration

AzureSearchConfiguration

DateTime

ISO 8601 combined date and time with time zone.

OAuthServerConfiguration

PgConnection

A connection to a PostgreSQL database.

Pool

An asynchronous pool of SQLx database connections.

Postgres

PostgreSQL database driver.

Url

A parsed URL record.

Utc

The UTC time zone. This is the most efficient time zone when you don’t need the local time. It is also used as an offset (which is also a dummy type).

UtilError

Error type used by all models. Used as the error type in UtilError, which is used by all the controllers in the application.

Uuid

A Universally Unique Identifier (UUID).

Enums

UtilErrorType

The type of UtilError that occured.

Traits

BackendError

The error types of this program all implement this trait for interoperability.

Connection

Represents a single database connection.

Context

Provides the context method for Result.

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

FromRow

A record that can be built from a row returned by the database.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Type

Indicates that a SQL type is supported for a database.

Type Aliases

PgPool

An alias for Pool, specialized for Postgres.

UtilResult

Used as the result types for all utils.

Derive Macros

Deserialize

FromRow

Serialize

Type