headless_lms_models

Module user_exercise_states

Structs

CourseExerciseMetrics

CourseUserPoints

CourseUserPointsInner

DatabaseUserCourseChapterExerciseProgress

EwusCourse

EwusExam

ExamUserPoints

ExamUserPointsInner

ExerciseUserCounts

ExerciseWithUserState

Convenience struct that combines user state to the exercise.

ExportedUserExerciseState

UserChapterMetrics

UserCourseChapterExerciseProgress

UserCourseMetrics

UserCourseProgress

UserExerciseState

UserExerciseStateUpdate

Enums

EwusContext

EwusCourseOrExam

ReviewingStage

Tells what stage of reviewing the user is currently in. Used for for peer review, self review, and manual review. If an exercise does not involve reviewing, the value of this stage will always be NotStarted.

Functions

get_all_for_user_and_course_or_exam

get_by_id

get_by_ids

get_course_metrics

get_course_metrics_indexed_by_module_id

get_course_metrics_open_chapters

get_course_users_counts_by_exercise

get_or_create_user_exercise_state

get_or_create_user_exercise_state_for_users

get_single_module_metrics

Gets course metrics for a single module.

get_user_course_chapter_exercises_progress

get_user_course_chapter_metrics

get_user_course_metrics

get_user_course_metrics_indexed_by_module_id

get_user_course_metrics_only_open_chapters

get_user_course_progress

get_user_exercise_state_if_exists

get_user_total_course_points

get_user_total_exam_points

Gets the total amount of points that the user has received from an exam.

get_users_current_by_exercise

merge_modules_with_metrics 🔒

stream_course_points

stream_exam_points

stream_user_exercise_states_for_course

update

TODO: should be moved to the user_exercise_state_updater as a private module so that this cannot be called outside of that module

update_exercise_progress

TODO: should be removed

update_reviewing_stage

upsert_selected_exercise_slide_id