headless_lms_models
Module user_exercise_task_states

Structs

UserExerciseTaskState

Functions

delete

figure_out_new_activity_progress

Returns a new state for the activity progress.

figure_out_new_grading_progress

Returns a new state for the grading progress.

figure_out_new_score_given

get

get_grading_summary_by_user_exercise_slide_state_id

insert

upsert_with_grading

Upserts user score from task grading results. The score can always increase or decrease, since they represent only a part of the whole user submission.

upsert_with_grading_status 🔒