Crate headless_lms_models

Functions and structs for interacting with the database.

Each submodule corresponds to a database table.

Re-exports

pub use self::error::ModelError;

pub use self::error::ModelErrorType;

pub use self::error::ModelResult;

Modules

chapters

course_background_question_answers

course_background_questions

course_exams

course_instance_enrollments

course_instances

course_language_groups

course_module_completion_registered_to_study_registries

course_module_completions

course_modules

courses

email_deliveries

email_templates

ended_processed_exams

error

Contains error and result types for all the model functions.

exams

exercise_repositories

exercise_service_info

exercise_services

exercise_slide_submissions

exercise_slides

exercise_task_gradings

exercise_task_regrading_submissions

exercise_task_submissions

exercise_tasks

exercises

feedback

file_uploads

glossary

library

material_references

open_university_registration_links

organizations

page_history

page_visit_datum

page_visit_datum_daily_visit_hashing_keys

pages

peer_review_configs

peer_review_question_submissions

peer_review_questions

peer_review_queue_entries

peer_review_submissions

pending_roles

playground_examples

prelude

Re-exports commonly used types for convenient use across the crate. Intended to be glob-imported like use crate::prelude::*;.

proposed_block_edits

proposed_page_edits

regradings

repository_exercises

roles

study_registry_registrars

teacher_grading_decisions

url_redirections

user_course_instance_exercise_service_variables

user_course_settings

user_exercise_slide_states

user_exercise_states

user_exercise_task_states

users

Enums

CourseOrExamId

Many database tables are related to either a course or an exam

PKeyPolicy

Helper struct to use with functions that insert data into the database.