Crate headless_lms_models

Functions and structs for interacting with the database.

Each submodule corresponds to a database table.

Re-exports

pub use self::error::ModelError;

pub use self::error::ModelErrorType;

pub use self::error::ModelResult;

Modules

certificate_configuration_to_requirements

certificate_configurations

certificate_fonts

chapters

chatbot_configurations

chatbot_conversation_messages

chatbot_conversations

chatbot_page_sync_statuses

code_giveaway_codes

code_giveaways

course_background_question_answers

course_background_questions

course_exams

course_instance_enrollments

course_instances

course_language_groups

course_module_completion_registered_to_study_registries

course_module_completions

course_modules

courses

email_deliveries

email_templates

ended_processed_exams

error

Contains error and result types for all the model functions.

exams

exercise_language_groups

exercise_repositories

exercise_service_info

exercise_services

exercise_slide_submissions

exercise_slides

exercise_task_gradings

exercise_task_regrading_submissions

exercise_task_submissions

exercise_tasks

exercises

feedback

file_uploads

generated_certificates

glossary

join_code_uses

library

material_references

offered_answers_to_peer_review_temporary

open_university_registration_links

organizations

other_domain_to_course_redirections

page_audio_files

page_history

page_language_groups

page_visit_datum

page_visit_datum_daily_visit_hashing_keys

page_visit_datum_summary_by_courses

page_visit_datum_summary_by_courses_countries

page_visit_datum_summary_by_courses_device_types

page_visit_datum_summary_by_pages

pages

partner_block

peer_or_self_review_configs

peer_or_self_review_question_submissions

peer_or_self_review_questions

peer_or_self_review_submissions

peer_review_queue_entries

pending_roles

playground_examples

prelude

Re-exports commonly used types for convenient use across the crate. Intended to be glob-imported like use crate::prelude::*;.

privacy_link

proposed_block_edits

proposed_page_edits

regradings

rejected_exercise_slide_submissions

repository_exercises

research_forms

roles

student_countries

study_registry_registrars

suspected_cheaters

teacher_grading_decisions

url_redirections

user_course_instance_exercise_service_variables

user_course_settings

user_details

user_exercise_slide_states

user_exercise_states

user_exercise_task_states

user_research_consents

users

Enums

CourseOrExamId

Many database tables are related to either a course or an exam

PKeyPolicy

Helper struct to use with functions that insert data into the database.

Traits

SpecFetcher

A “trait alias” so this for<'a> … string doesn’t need to be repeated everywhere Arguments: Url: The URL that the request is sent to (the exercise service’s endpoint) &str: Exercise type/service slug Option<Value>: The Json for the request, for example the private spec in a public spec request