Module headless_lms_models::certificate_configurations

Structs

CertificateConfiguration

CertificateConfigurationAndRequirements

DatabaseCertificateConfiguration

DatabaseCertificateConfigurationInner 🔒

Enums

CertificateTextAnchor

How text should be positioned relative to the given coordinates. See https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/text-anchor.

PaperSize

Functions

delete

get_all_certifcate_configurations_requiring_only_one_module_and_no_course_instance

Finds all configurations that applies to any instance of a course module and requires only one course module. *

get_by_id

get_default_certificate_configurations_and_requirements_by_course_instance

A default certificate configuration requires only one course module.

get_default_configuration_by_course_module_and_course_instance

get_required_course_instance_ids

insert

update