Module headless_lms_models::course_module_completions

Structs

CourseModuleCompletion

CourseModuleCompletionWithRegistrationInfo

NewCourseModuleCompletion

StudyRegistryCompletion

Completion in the form that is recognized by authorized third party study registry registrars.

StudyRegistryGrade

Grading object that maps the system grading information to Sisu’s grading scales.

Enums

CourseModuleCompletionGranter

Functions

delete

get_all_by_course_instance_and_user_id

Gets all module completions for the user on a single course instance. There can be multiple modules in a single course, so the result is a Vec.

get_all_by_course_instance_id

get_all_by_course_module_instance_and_user_ids

get_all_by_user_id

get_all_with_registration_information_by_course_instance_id

Gets summaries for all completions on the given course instance.

get_automatic_completion_by_course_module_instance_and_user_ids

Gets automatically granted course module completion for the given user on the specified course instance. This entry is quaranteed to be unique in database by the index course_module_automatic_completion_uniqueness.

get_by_id

get_by_ids

get_by_ids_as_map

get_count_of_distinct_completors_by_course_id

Get the number of students that have completed the course

get_latest_by_course_module_instance_and_user_ids

Gets latest created completion for the given user on the specified course instance.

insert

stream_by_course_module_id

Streams completions.

update_completion_registration_attempt_date

update_prerequisite_modules_completed

user_has_completed_course_module_on_instance

Checks whether the user has any completions for the given course module on the specified course instance.