Module headless_lms_models::course_module_completions

Structs

CourseModuleCompletion

NewCourseModuleCompletion

StudyRegistryCompletion

Completion in the form that is recognized by authorized third party study registry registrars.

StudyRegistryGrade

Grading object that maps the system grading information to Sisu’s grading scales.

Functions

delete

get_by_course_instance_and_user_ids

Gets all module completions for the user on a single course instance. There can be multiple modules in a single course, so the result is a Vec.

get_by_course_module_instance_and_user_ids

get_by_id

get_by_ids

get_by_ids_as_map

insert

stream_by_course_module_id

update_completion_registration_attempt_date

update_prerequisite_modules_completed