Module headless_lms_models::course_modules

Structs

AutomaticCompletionRequirements

CourseModule

Based on CourseModulesSchema but completion_policy parsed and addded (and some not needeed fields removed).

CourseModulesSchema 🔒

Matches the columns in the database.

ModifiedModule

ModuleUpdates

NewCourseModule

NewModule

Record 🔒

Record 🔒

Enums

CompletionPolicy

Functions

delete

get_all_uh_course_codes_for_open_university

get_by_course_id

get_by_course_id_as_map

Gets course modules for the given course as a map, indexed by the id field.

get_by_exercise_id

Gets course module where the given exercise belongs to. This will result in an error in the case of an exam exercise.

get_by_id

get_default_by_course_id

get_ids_by_course_slug_or_uh_course_code

Gets all course modules with a matching uh_course_code or course slug.

insert

rename

update

update_automatic_completion_status

update_certification_enabled

update_enable_registering_completion_to_uh_open_university

update_modules

update_uh_course_code

update_with_order_number