headless_lms_models

Module email_deliveries

Structs

CheckUserAndTemplateRow 🔒

Email

EmailDelivery

EmailDeliveryError

EmailDeliveryErrorInsert

Constants

FETCH_LIMIT

Functions

fetch_emails

increment_retry_and_mark_non_retryable

increment_retry_and_schedule

insert_email_delivery

Inserts an email delivery; fails if the user or email template is soft-deleted.

insert_email_delivery_error

mark_as_sent

soft_delete_unsent_retryable_deliveries_for_user

Soft-deletes unsent, still-retryable email deliveries for a user. Call when soft-deleting the user so pending deliveries are not retried.