Module headless_lms_models::exercise_service_info

Structs

CourseMaterialExerciseServiceInfo

ExerciseServiceInfo

ExerciseServiceInfoApi

PathInfo

Functions

fetch_and_upsert_service_info

fetch_service_info

get_all_exercise_services_by_type

get_course_material_service_info_by_exercise_type

Returns service info meant for the course material. If no service info is found and fetching it fails, we return None to indicate that the service info is unavailable.

get_selected_exercise_services_by_type

get_service_info

get_service_info_by_exercise_service

get_service_info_by_exercise_type

insert

upsert_service_info