Module headless_lms_models::exercise_services

Structs

ExerciseService

ExerciseServiceIframeRenderingInfo

Exercise service definition that the CMS can use to render the editor view.

ExerciseServiceNewOrUpdate

Functions

delete_exercise_service

get_all_exercise_services_iframe_rendering_infos

get_exercise_service

get_exercise_service_by_exercise_type

get_exercise_service_externally_preferred_baseurl

get_exercise_service_internally_preferred_baseurl

get_exercise_service_internally_preferred_baseurl_by_exercise_type

get_exercise_services

get_internal_grade_url

Returns a url that can be used to grade a submission for this exercise service.

get_internal_public_spec_url

Returns a url that can be used to generate a public version of a private spec.

get_model_solution_url

insert_exercise_service

update_exercise_service