headless_lms_models

Module exercise_task_submissions

Structs

ExerciseTaskSubmission

ExportedCourseSubmission

ExportedSubmission

PeerOrSelfReviewsReceived

SubmissionData

Functions

get_by_exercise_slide_submission_id

get_by_id

get_course_and_exam_id

get_exercise_task_submission_info_by_exercise_slide_submission_id

Used to get the necessary info for rendering a submission either when we’re viewing a submission, or we’re conducting a peer review.

get_ids_by_exercise_id

get_latest_submission_ids_by_exercise_id

Similar to get_ids_by_exercise_id but returns the record with the highest created_at for a user_id

get_peer_reviews_received

get_submission

get_user_custom_view_exercise_tasks_by_module_and_exercise_type

get_user_latest_exercise_task_submissions_by_course_module_and_exercise_type

get all submissions for user and course module and exercise type

get_users_latest_exercise_task_submissions_for_exercise_slide

insert

insert_with_id

set_grading_id

stream_course_submissions

stream_exam_submissions