Module headless_lms_models::library::grading

Structs

ExerciseSlideSubmissionWithTasks

ExerciseStateUpdateNeedToUpdatePeerReviewStatusWithThis

If passed to to an exercise state update, it will update the peer review status with the given information

StudentExerciseSlideSubmission

Contains data sent by the student when they make a submission for an exercise slide.

StudentExerciseSlideSubmissionResult

StudentExerciseTaskSubmission

StudentExerciseTaskSubmissionResult

Enums

GradingPolicy

Functions

calculate_average_received_peer_review_score ğŸ”’

create_fixed_grading_for_submission_task ğŸ”’

create_user_exercise_slide_submission

Inserts user submission to database. Tasks within submission are validated to make sure that they belong to the correct exercise slide.

derive_new_user_exercise_state ğŸ”’

derive_score_given_when_peer_review_complete ğŸ”’

grade_user_submission

grade_user_submission_task ğŸ”’

propagate_user_exercise_state_update_from_exercise_task_grading_result

Updates the user exercise state starting from a single task, and propagates the update up to the whole user exercise state.

propagate_user_exercise_state_update_from_slide ğŸ”’

Updates the user exercise state starting from a slide state, and propagates the update up to the whole user exercise state.

update_grading_with_single_regrading_result

update_user_exercise_slide_state ğŸ”’

update_user_exercise_state ğŸ”’

update_user_exercise_state_peer_review_status