Module headless_lms_models::library::grading

Collection of functions used for processing and evaluating user submissions for exercises.

Structs

AnswerRequiringAttentionWithTasks

AnswersRequiringAttention

ExerciseSlideSubmissionWithTasks

ExerciseStateUpdateNeedToUpdatePeerReviewStatusWithThis

If passed to to an exercise state update, it will update the peer review status with the given information

StudentExerciseSlideSubmission

Contains data sent by the student when they make a submission for an exercise slide.

StudentExerciseSlideSubmissionResult

StudentExerciseTaskSubmission

StudentExerciseTaskSubmissionResult

Enums

GradingPolicy

Functions

create_fixed_grading_for_submission_task 🔒

create_user_exercise_slide_submission

Inserts user submission to database. Tasks within submission are validated to make sure that they belong to the correct exercise slide.

get_paginated_answers_requiring_attention_for_exercise

Gets submissions that require input from the teacher to continue processing.

grade_user_submission

grade_user_submission_task 🔒

propagate_user_exercise_state_update_from_exercise_task_grading_result

Updates the user exercise state starting from a single task, and propagates the update up to the whole user exercise state.

update_grading_with_single_regrading_result

update_user_exercise_slide_state 🔒

update_user_exercise_slide_state_and_user_exercise_state 🔒

Updates the user exercise state starting from a slide state, and propagates the update up to the whole user exercise state.