Module headless_lms_models::library::progressing

Structs

CompletionRegistrationLink

UserCompletionInformation

UserModuleCompletionStatus

Functions

create_course_module_completion_if_eligible 🔒

Creates completion for the user if eligible and previous one doesn’t exist. Returns a boolean indicating whether a completion exists after calling this function.

get_completion_registration_link_and_save_attempt

get_user_completion_information

get_user_module_completion_statuses_for_course_instance

Gets course modules with user’s completion status for the given instance.

prerequisite_modules_are_completed 🔒

process_all_course_instance_completions

update_automatic_completion_status_and_grant_if_eligible

Checks whether the course module can be completed automatically and creates an entry for completion if the user meets the criteria. Also re-checks module completion prerequisites if the module is completed.

update_module_completion_prerequisite_statuses 🔒

user_has_completed_course_module 🔒

user_is_eligible_for_automatic_completion 🔒