Module headless_lms_models::library::user_exercise_state_updater::data_loader

Functions

load_current_user_exercise_state ğŸ”’

load_exercise ğŸ”’

load_given_peer_review_submissions ğŸ”’

load_latest_exercise_slide_submission ğŸ”’

load_latest_exercise_slide_submission_received_peer_review_question_submissions ğŸ”’

load_latest_teacher_grading_decision ğŸ”’

load_peer_review_config ğŸ”’

load_peer_review_information ğŸ”’

load_peer_review_queue_entry ğŸ”’

load_required_data ğŸ”’

Returns an object with all dependencies for the user_exercise_state update loaded. Either uses a preloaded value from already_loaded_internal_dependencies or fetches the necessary information from the database.

load_user_exercise_slide_state_grading_summary ğŸ”’