Module headless_lms_models::marketing_consents

Structs

MarketingMailingListAccessToken

UserEmailSubscription

UserMarketingConsent

UserMarketingConsentWithDetails

Functions

fetch_all_marketing_mailing_list_access_tokens

fetch_all_unsynced_updated_emails

Fetches email, email subscription status and user ID for users whose details have been updated after their marketing consent was last synced to Mailchimp

fetch_all_unsynced_user_marketing_consents_by_course_language_group_id

Fetches all user marketing consents with detailed user information for a specific course language group, if they haven’t been synced to Mailchimp or if there have been updates since the last sync.

fetch_user_marketing_consent

update_synced_to_mailchimp_at_to_all_synced_users

Used to update the synced_to_mailchimp_at to a list of users when they are successfully synced to mailchimp

update_unsubscribed_users_from_mailchimp_in_bulk

Updates user consents to false in bulk using Mailchimp data.

update_user_mailchimp_id_at_to_all_synced_users

Used to add the user_mailchimp_ids to a list of new users when they are successfully synced to mailchimp

upsert_marketing_consent