Module headless_lms_models::peer_or_self_review_question_submissions

Structs

PeerOrSelfReviewQuestionAndAnswer

PeerOrSelfReviewQuestionSubmission

PeerReviewWithQuestionsAndAnswers

Enums

PeerOrSelfReviewAnswer

Functions

bundle_peer_or_self_review_questions_and_answers 🔒

Groups answers to peer reviews by peer review ids.

get_by_peer_reviews_question_ids

get_given_peer_reviews

Returns the peer reviews given by this user. Does not return the self reviews given by the user.

get_question_submissions_from_from_peer_or_self_review_submission_ids

get_questions_and_answers_by_submission_id

get_received_question_submissions_for_exercise_slide_submission

insert