Module headless_lms_models::peer_review_configs

Structs

CmsPeerReviewConfig

CmsPeerReviewConfiguration

PeerReviewConfig

Enums

PeerReviewAcceptingStrategy

Determines how we will treat the answer being peer reviewed once it has received enough reviews and the student has given enough peer reviews.

Functions

delete

delete_peer_reviews_by_exrcise_ids

get_by_exercise_or_course_id

get_by_id

get_cms_peer_review_by_id

get_course_default_cms_peer_review

get_course_material_peer_review_data

get_default_for_course_by_course_id

get_peer_reviews_by_page_id

insert

insert_with_id

try_to_get_by_exercise_id

upsert_course_default_cms_peer_review_and_questions

upsert_with_id