Module headless_lms_models::peer_review_queue_entries

Structs

PeerReviewQueueEntry

Functions

get_all_that_need_peer_reviews_by_exercise_id

Gets multiple records of PeerReviewQueueEntry that still require more peer reviews, ordered by peer review priority.

get_by_id

get_by_receiving_peer_reviews_submission_and_course_instance_ids 🔒

get_by_user_and_exercise_and_course_instance_ids

get_entries_that_need_reviews_and_are_older_than

get_many_by_exercise_id_and_review_priority

Gets multiple records of PeerReviewQueueEntry ordered by peer review priority. Also returns entries that don’t need peer review.

get_many_that_need_peer_reviews_by_exercise_id_and_review_priority

Gets multiple records of PeerReviewQueueEntry that still require more peer reviews, ordered by peer review priority.

increment_peer_review_priority

insert

remove_from_queue 🔒

remove_from_queue_and_add_to_manual_review

remove_from_queue_and_give_full_points

remove_queue_entries_for_unusual_reason

try_to_get_by_receiving_submission_and_course_instance_ids

try_to_get_by_user_and_exercise_and_course_instance_ids

update

update_received_enough_peer_reviews

upsert_peer_review_priority

Inserts or updates the queue entry indexed by user_id, exercise_id and course_instance_id.