Module headless_lms_models::prelude

Re-exports commonly used types for convenient use across the crate. Intended to be glob-imported like use crate::prelude::*;.

Re-exports

pub use crate::error::TryToOptional;

pub use crate::CourseOrExamId;

pub use crate::ModelError;

pub use crate::ModelErrorType;

pub use crate::ModelResult;

pub use crate::PKeyPolicy;

Structs

DateTime

ISO 8601 combined date and time with time zone.

Pagination

Represents the URL query parameters page and limit, used for paginating database queries.

PgConnection

A connection to a PostgreSQL database.

Utc

The UTC time zone. This is the most efficient time zone when you don’t need the local time. It is also used as an offset (which is also a dummy type).

Uuid

A Universally Unique Identifier (UUID).

Traits

BackendError

The error types of this program all implement this trait for interoperability.

Connection

Represents a single database connection.

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

FromRow

A record that can be built from a row returned by the database.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Type

Indicates that a SQL type is supported for a database.

Derive Macros

Deserialize

FromRow

Serialize

Type