Module headless_lms_models::user_exercise_states

Structs

CourseInstanceExerciseMetrics

CourseInstanceUserPoints

CourseInstanceUserPointsInner

DatabaseUserCourseInstanceChapterExerciseProgress

EwusCourse

EwusExam

ExamUserPoints

ExamUserPointsInner

ExerciseUserCounts

ExerciseWithUserState

Convenience struct that combines user state to the exercise.

ExportedUserExerciseState

UserChapterMetrics

UserCourseInstanceChapterExerciseProgress

UserCourseInstanceMetrics

UserCourseInstanceProgress

UserExerciseState

UserExerciseStateUpdate

Enums

CourseInstanceOrExamId

Either a course instance or exam id.

EwusContext

EwusCourseOrExam

ReviewingStage

Tells what stage of reviewing the user is currently in. Used for for peer review, self review, and manual review. If an exercise does not involve reviewing, the value of this stage will always be NotStarted.

Functions

get_all_for_user_and_course_instance_or_exam

get_by_id

get_by_ids

get_course_instance_metrics

get_course_instance_metrics_indexed_by_module_id

get_course_users_counts_by_exercise

get_or_create_user_exercise_state

get_or_create_user_exercise_state_for_users

get_single_module_course_instance_metrics

Gets course instance metrics for a single module.

get_user_course_instance_chapter_exercises_progress

get_user_course_instance_chapter_metrics

get_user_course_instance_metrics

get_user_course_instance_metrics_indexed_by_module_id

get_user_course_instance_progress

get_user_exercise_state_if_exists

get_user_total_course_points

get_user_total_exam_points

Gets the total amount of points that the user has received from an exam.

get_users_current_by_exercise

merge_modules_with_metrics 🔒

stream_course_instance_points

stream_exam_points

stream_user_exercise_states_for_course

update

TODO: should be moved to the user_exercise_state_updater as a private module so that this cannot be called outside of that module

update_exercise_progress

TODO: should be removed

update_reviewing_stage

upsert_selected_exercise_slide_id